+++ sdk/storage/storage-blob/CHANGELOG.md
@@ -11,6 +11,7 @@
   - Added BlobClient.setLegalHold()
 - Added support for listing deleted root blobs with versions `ContainerClient.listBlobFlat()` and `ContainerClient.listBlobHierarchy()`.
 - Added support for OAuth copy sources for synchronous copy operations.
+- Added support for Parquet as an input format in `BlockBlobClient.query()`.
 - With the dropping of support for Node.js versions that are no longer in LTS, the dependency on `@types/node` has been updated to version 12. Read our [support policy](https://github.com/Azure/azure-sdk-for-js/blob/main/SUPPORT.md) for more details.
 - Changed TS compilation target to ES2017 in order to produce smaller bundles and use more native platform features